ingestion/requirements.txt
boto3>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
pyarrow>=14.0.0
yfinance>=0.2.36
numpy>=1.26.0